backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
shap==0.44.1
google-generativeai==0.4.1
google-cloud-storage==2.14.0
google-cloud-language==2.13.0
firebase-admin==6.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.6.0
httpx==0.26.0
scipy==1.12.0
